+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ### 12.65.0
 
+* `[fmtutil]` Added method `Align` for aligning text with ANSI control sequences (_for example colors_)
+* `[usage]` Added feature for adding and rendering environment info
+
+### 12.64.1
+
 * `[processes]` `ProcessInfo.Childs` renamed to `ProcessInfo.Children`
 * Fixed typos